Cleaned up old files and fixed naming etc.
