--- daemon/HWFmessage.fbs
@@ -1,9 +1,0 @@
-namespace Message;
-
-table Message{
-    agent_id:string;
-    cmd:[string];
-    data:[byte];
-}
-
-root_type Message;
