--- daemon/schema.fbs
@@ -1,7 +1,0 @@
-namespace HelloWorld;
-
-table HelloWorld {
-    msg:string;
-}
-
-root_type HelloWorld;
