+++ server/src/hwfSchema_generated.ts
@@ -1,6 +1,18 @@
 // automatically generated by the FlatBuffers compiler, do not modify
 
-import flatbuffers from "flatbuffers"
+import  flatbuffers  from "flatbuffers"
+/**
+ * @enum {number}
+ */
+export namespace schema{
+export enum MessageBody{
+  NONE= 0,
+  Task= 1,
+  GetResult= 2,
+  GetHardwarePool= 3,
+  File= 4
+}};
+
 /**
  * @constructor
  */
@@ -38,37 +50,27 @@ type():number {
 };
 
 /**
- * @param schema.Task= obj
- * @returns schema.Task|null
+ * @returns schema.MessageBody
  */
-task(obj?:schema.Task):schema.Task|null {
+bodyType():schema.MessageBody {
   var offset = this.bb!.__offset(this.bb_pos, 6);
-  return offset ? (obj || new schema.Task).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? /**  */ (this.bb!.readUint8(this.bb_pos + offset)) : schema.MessageBody.NONE;
 };
 
 /**
- * @param schema.GetResult= obj
- * @returns schema.GetResult|null
+ * @param flatbuffers.Table obj
+ * @returns ?flatbuffers.Table
  */
-getResult(obj?:schema.GetResult):schema.GetResult|null {
+body<T extends flatbuffers.Table>(obj:T|flatbuffers.Table):T|null|flatbuffers.Table {
   var offset = this.bb!.__offset(this.bb_pos, 8);
-  return offset ? (obj || new schema.GetResult).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
-};
-
-/**
- * @param schema.GetHardwarePool= obj
- * @returns schema.GetHardwarePool|null
- */
-getHardwarePool(obj?:schema.GetHardwarePool):schema.GetHardwarePool|null {
-  var offset = this.bb!.__offset(this.bb_pos, 10);
-  return offset ? (obj || new schema.GetHardwarePool).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+  return offset ? this.bb!.__union(obj, this.bb_pos + offset) : null;
 };
 
 /**
  * @param flatbuffers.Builder builder
  */
 static startMessage(builder:flatbuffers.Builder) {
-  builder.startObject(4);
+  builder.startObject(3);
 };
 
 /**
@@ -81,26 +83,18 @@ static addType(builder:flatbuffers.Builder, type:number) {
 
 /**
  * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset taskOffset
+ * @param schema.MessageBody bodyType
  */
-static addTask(builder:flatbuffers.Builder, taskOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, taskOffset, 0);
+static addBodyType(builder:flatbuffers.Builder, bodyType:schema.MessageBody) {
+  builder.addFieldInt8(1, bodyType, schema.MessageBody.NONE);
 };
 
 /**
  * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset getResultOffset
+ * @param flatbuffers.Offset bodyOffset
  */
-static addGetResult(builder:flatbuffers.Builder, getResultOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(2, getResultOffset, 0);
-};
-
-/**
- * @param flatbuffers.Builder builder
- * @param flatbuffers.Offset getHardwarePoolOffset
- */
-static addGetHardwarePool(builder:flatbuffers.Builder, getHardwarePoolOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(3, getHardwarePoolOffset, 0);
+static addBody(builder:flatbuffers.Builder, bodyOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(2, bodyOffset, 0);
 };
 
 /**
@@ -120,12 +114,11 @@ static finishMessageBuffer(builder:flatbuffers.Builder, offset:flatbuffers.Offse
   builder.finish(offset);
 };
 
-static createMessage(builder:flatbuffers.Builder, type:number, taskOffset:flatbuffers.Offset, getResultOffset:flatbuffers.Offset, getHardwarePoolOffset:flatbuffers.Offset):flatbuffers.Offset {
+static createMessage(builder:flatbuffers.Builder, type:number, bodyType:schema.MessageBody, bodyOffset:flatbuffers.Offset):flatbuffers.Offset {
   Message.startMessage(builder);
   Message.addType(builder, type);
-  Message.addTask(builder, taskOffset);
-  Message.addGetResult(builder, getResultOffset);
-  Message.addGetHardwarePool(builder, getHardwarePoolOffset);
+  Message.addBodyType(builder, bodyType);
+  Message.addBody(builder, bodyOffset);
   return Message.endMessage(builder);
 }
 }
@@ -422,12 +415,21 @@ static getRootAsTask(bb:flatbuffers.ByteBuffer, obj?:Task):Task {
 };
 
 /**
+ * @param schema.Hardware= obj
+ * @returns schema.Hardware|null
+ */
+hardware(obj?:schema.Hardware):schema.Hardware|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? (obj || new schema.Hardware).__init(this.bb!.__indirect(this.bb_pos + offset), this.bb!) : null;
+};
+
+/**
  * @param number index
  * @param schema.Stage= obj
  * @returns schema.Stage
  */
 stages(index: number, obj?:schema.Stage):schema.Stage|null {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
+  var offset = this.bb!.__offset(this.bb_pos, 6);
   return offset ? (obj || new schema.Stage).__init(this.bb!.__indirect(this.bb!.__vector(this.bb_pos + offset) + index * 4), this.bb!) : null;
 };
 
@@ -435,7 +437,7 @@ stages(index: number, obj?:schema.Stage):schema.Stage|null {
  * @returns number
  */
 stagesLength():number {
-  var offset = this.bb!.__offset(this.bb_pos, 4);
+  var offset = this.bb!.__offset(this.bb_pos, 6);
   return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
 };
 
@@ -447,7 +449,7 @@ stagesLength():number {
 artifacts(index: number):string
 artifacts(index: number,optionalEncoding:flatbuffers.Encoding):string|Uint8Array
 artifacts(index: number,optionalEncoding?:any):string|Uint8Array|null {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
+  var offset = this.bb!.__offset(this.bb_pos, 8);
   return offset ? this.bb!.__string(this.bb!.__vector(this.bb_pos + offset) + index * 4, optionalEncoding) : null;
 };
 
@@ -455,7 +457,7 @@ artifacts(index: number,optionalEncoding?:any):string|Uint8Array|null {
  * @returns number
  */
 artifactsLength():number {
-  var offset = this.bb!.__offset(this.bb_pos, 6);
+  var offset = this.bb!.__offset(this.bb_pos, 8);
   return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
 };
 
@@ -463,7 +465,15 @@ artifactsLength():number {
  * @param flatbuffers.Builder builder
  */
 static startTask(builder:flatbuffers.Builder) {
-  builder.startObject(2);
+  builder.startObject(3);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset hardwareOffset
+ */
+static addHardware(builder:flatbuffers.Builder, hardwareOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, hardwareOffset, 0);
 };
 
 /**
@@ -471,7 +481,7 @@ static startTask(builder:flatbuffers.Builder) {
  * @param flatbuffers.Offset stagesOffset
  */
 static addStages(builder:flatbuffers.Builder, stagesOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(0, stagesOffset, 0);
+  builder.addFieldOffset(1, stagesOffset, 0);
 };
 
 /**
@@ -500,7 +510,7 @@ static startStagesVector(builder:flatbuffers.Builder, numElems:number) {
  * @param flatbuffers.Offset artifactsOffset
  */
 static addArtifacts(builder:flatbuffers.Builder, artifactsOffset:flatbuffers.Offset) {
-  builder.addFieldOffset(1, artifactsOffset, 0);
+  builder.addFieldOffset(2, artifactsOffset, 0);
 };
 
 /**
@@ -533,8 +543,9 @@ static endTask(builder:flatbuffers.Builder):flatbuffers.Offset {
   return offset;
 };
 
-static createTask(builder:flatbuffers.Builder, stagesOffset:flatbuffers.Offset, artifactsOffset:flatbuffers.Offset):flatbuffers.Offset {
+static createTask(builder:flatbuffers.Builder, hardwareOffset:flatbuffers.Offset, stagesOffset:flatbuffers.Offset, artifactsOffset:flatbuffers.Offset):flatbuffers.Offset {
   Task.startTask(builder);
+  Task.addHardware(builder, hardwareOffset);
   Task.addStages(builder, stagesOffset);
   Task.addArtifacts(builder, artifactsOffset);
   return Task.endTask(builder);
@@ -738,6 +749,275 @@ static createGetHardwarePool(builder:flatbuffers.Builder, hardwareOffset:flatbuf
   GetHardwarePool.startGetHardwarePool(builder);
   GetHardwarePool.addHardware(builder, hardwareOffset);
   return GetHardwarePool.endGetHardwarePool(builder);
+}
+}
+}
+/**
+ * @constructor
+ */
+export namespace schema{
+export class File {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns File
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):File {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param File= obj
+ * @returns File
+ */
+static getRootAsFile(bb:flatbuffers.ByteBuffer, obj?:File):File {
+  return (obj || new File).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+filename():string|null
+filename(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+filename(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @returns number
+ */
+packetnumber():number {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.readInt32(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns boolean
+ */
+eof():boolean {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? !!this.bb!.readInt8(this.bb_pos + offset) : false;
+};
+
+/**
+ * @param number index
+ * @returns number
+ */
+data(index: number):number|null {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? this.bb!.readUint8(this.bb!.__vector(this.bb_pos + offset) + index) : 0;
+};
+
+/**
+ * @returns number
+ */
+dataLength():number {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? this.bb!.__vector_len(this.bb_pos + offset) : 0;
+};
+
+/**
+ * @returns Uint8Array
+ */
+dataArray():Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 10);
+  return offset ? new Uint8Array(this.bb!.bytes().buffer, this.bb!.bytes().byteOffset + this.bb!.__vector(this.bb_pos + offset), this.bb!.__vector_len(this.bb_pos + offset)) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startFile(builder:flatbuffers.Builder) {
+  builder.startObject(4);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset filenameOffset
+ */
+static addFilename(builder:flatbuffers.Builder, filenameOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, filenameOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number packetnumber
+ */
+static addPacketnumber(builder:flatbuffers.Builder, packetnumber:number) {
+  builder.addFieldInt32(1, packetnumber, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param boolean eof
+ */
+static addEof(builder:flatbuffers.Builder, eof:boolean) {
+  builder.addFieldInt8(2, +eof, +false);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset dataOffset
+ */
+static addData(builder:flatbuffers.Builder, dataOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(3, dataOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param Array.<number> data
+ * @returns flatbuffers.Offset
+ */
+static createDataVector(builder:flatbuffers.Builder, data:number[] | Uint8Array):flatbuffers.Offset {
+  builder.startVector(1, data.length, 1);
+  for (var i = data.length - 1; i >= 0; i--) {
+    builder.addInt8(data[i]);
+  }
+  return builder.endVector();
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param number numElems
+ */
+static startDataVector(builder:flatbuffers.Builder, numElems:number) {
+  builder.startVector(1, numElems, 1);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endFile(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+static createFile(builder:flatbuffers.Builder, filenameOffset:flatbuffers.Offset, packetnumber:number, eof:boolean, dataOffset:flatbuffers.Offset):flatbuffers.Offset {
+  File.startFile(builder);
+  File.addFilename(builder, filenameOffset);
+  File.addPacketnumber(builder, packetnumber);
+  File.addEof(builder, eof);
+  File.addData(builder, dataOffset);
+  return File.endFile(builder);
+}
+}
+}
+/**
+ * @constructor
+ */
+export namespace schema{
+export class Hardware {
+  bb: flatbuffers.ByteBuffer|null = null;
+
+  bb_pos:number = 0;
+/**
+ * @param number i
+ * @param flatbuffers.ByteBuffer bb
+ * @returns Hardware
+ */
+__init(i:number, bb:flatbuffers.ByteBuffer):Hardware {
+  this.bb_pos = i;
+  this.bb = bb;
+  return this;
+};
+
+/**
+ * @param flatbuffers.ByteBuffer bb
+ * @param Hardware= obj
+ * @returns Hardware
+ */
+static getRootAsHardware(bb:flatbuffers.ByteBuffer, obj?:Hardware):Hardware {
+  return (obj || new Hardware).__init(bb.readInt32(bb.position()) + bb.position(), bb);
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+cpu():string|null
+cpu(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+cpu(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 4);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+gpu():string|null
+gpu(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+gpu(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 6);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Encoding= optionalEncoding
+ * @returns string|Uint8Array|null
+ */
+os():string|null
+os(optionalEncoding:flatbuffers.Encoding):string|Uint8Array|null
+os(optionalEncoding?:any):string|Uint8Array|null {
+  var offset = this.bb!.__offset(this.bb_pos, 8);
+  return offset ? this.bb!.__string(this.bb_pos + offset, optionalEncoding) : null;
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ */
+static startHardware(builder:flatbuffers.Builder) {
+  builder.startObject(3);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset cpuOffset
+ */
+static addCpu(builder:flatbuffers.Builder, cpuOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(0, cpuOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset gpuOffset
+ */
+static addGpu(builder:flatbuffers.Builder, gpuOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(1, gpuOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @param flatbuffers.Offset osOffset
+ */
+static addOs(builder:flatbuffers.Builder, osOffset:flatbuffers.Offset) {
+  builder.addFieldOffset(2, osOffset, 0);
+};
+
+/**
+ * @param flatbuffers.Builder builder
+ * @returns flatbuffers.Offset
+ */
+static endHardware(builder:flatbuffers.Builder):flatbuffers.Offset {
+  var offset = builder.endObject();
+  return offset;
+};
+
+static createHardware(builder:flatbuffers.Builder, cpuOffset:flatbuffers.Offset, gpuOffset:flatbuffers.Offset, osOffset:flatbuffers.Offset):flatbuffers.Offset {
+  Hardware.startHardware(builder);
+  Hardware.addCpu(builder, cpuOffset);
+  Hardware.addGpu(builder, gpuOffset);
+  Hardware.addOs(builder, osOffset);
+  return Hardware.endHardware(builder);
 }
 }
 }
