--- testSchema.fbs
@@ -1,50 +1,0 @@
-namespace schema;
-
-union MessageBody {Task, GetResult, GetHardwarePool, File}
-
-table Message {
-    type:int;
-    body:MessageBody;
-}
-
-table Stage {
-    name:string;
-    data:[byte];
-    cmd_list:[string];
-    track_time:bool = true;
-    track_ram:bool = false;
-    track_cpu:bool = false;
-    track_gpu:bool = false;
-    comment:string;
-}
-
-table Task {
-    hardware:Hardware;
-    stages:[Stage]; //Be sure to create a vector that can store several Stages
-    artifacts:[string];
-}
-
-table GetResult {
-    id_list:[string];
-}
-
-table GetHardwarePool {
-    hardware:[string];
-}
-
-table File {
-    filename:string; 
-    packetnumber:int; 
-    eof:bool;
-    data:[ubyte];
-}
-
-table Hardware {
-    cpu:string;
-    gpu:string;
-    os:string;
-}
-
-root_type Message;
-
-
